requirements-dev.txt
# Requirements para desenvolvimento (Windows)
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
authlib==1.3.0
requests-oauthlib==1.3.1
httpx==0.25.2
python-dotenv==1.0.0
starlette==0.27.0
itsdangerous==2.1.2
# Para Windows - usar uma dessas opções:
# psycopg2==2.9.9
# OU usar SQLite para desenvolvimento local:
# (sem psycopg2, o SQLite será usado automaticamente)
supabase==2.3.4
